Added a dropzone
